dcbot/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
python-dotenv==1.0.1
setuptools==69.0.3
websocket-client==1.7.0
websockets==12.0
wheel==0.42.0
yarl==1.9.4
